feat: user internal post and refactor code
