+++ backend/game/gameapp/match_objects/__init__.py
@@ -1,0 +1,5 @@
+from .match import Match, match_decided
+from .matchdict import match_dict
+from .matchuser import MatchUser
+from .process import MatchProcess
+from .timeout import WaitingProcess
