+++ backend/game/gameapp/match_objects/matchuser.py
@@ -1,0 +1,7 @@
+from typing import TypedDict
+
+
+class MatchUser(TypedDict):
+    id: int
+    name: str
+    sid: str
